web/requirements.txt
amqp==2.5.2
beautifulsoup4==4.8.1
billiard==3.6.1.0
bs4==0.0.1
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
djangorestframework==3.10.3
idna==2.8
importlib-metadata==1.0.0
kombu==4.6.6
more-itertools==8.0.0
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
psycopg2-binary==2.8.4
gunicorn
redis
